# docs/apifox-openapi.yaml
openapi: 3.0.3
info:
  title: X-Pilot Agentic Server API
  version: 1.0.0
  description: |
    可直接导入 Apifox 的 OpenAPI 文档。
    注意：
    - 认证方式：`Authorization: Bearer <apiKey>`
    - 业务文稿接口路径是 `/api/business/document-generation/*`，不是 `document-generator`
servers:
  - url: http://192.168.50.54:7800
    description: Local Network Dev
security:
  - BearerAuth: []
...
paths:
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: 查询任务列表/统计
      parameters:
        - in: query
          name: agent
          required: false
          schema: { type: string }
          example: document-generator
        - in: query
          name: parent
          required: false
          schema: { type: string }
          example: task_parent_1
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                stats:
                  pending: 1
                  running: 0
                  completed: 12
                  failed: 1
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: 查询任务详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          example: task_abc123
      responses:
        "200":
          description: 成功
          content:
            application/json:
              example:
                id: task_abc123
                status: completed
                agent: document-generator
    delete:
      tags: [Tasks]
      summary: 取消任务
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
          example: task_abc123
      responses:
        "200":
          description: 取消成功
          content:
            application/json:
              example: { success: true }
        "400":
          description: 无法取消
          content:
            application/json:
              example:
                error: "Task cannot be cancelled (not pending or not found)"
...
  /v1/conversations:
    get:
      tags: [Dify-Compat]
      summary: 获取会话列表
      description: 获取当前用户的所有会话列表（Dify 兼容接口，需 STREAM_PROTOCOL=dify）
      parameters:
        - name: user
          in: query
          schema:
            type: string
          description: 用户标识（由 auth 中间件解析，此参数可选）
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: first_id
          in: query
          schema:
            type: string
          description: 游标分页 - 上一页最后一个会话 ID
      responses:
        "200":
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  limit:
                    type: integer
                  has_more:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        inputs:
                          type: object
                        status:
                          type: string
                        introduction:
                          type: string
                        created_at:
                          type: number
                        updated_at:
                          type: number
              example:
                limit: 20
                has_more: false
                data:
                  - id: "test0321"
                    name: "电子电工原理及故障诊断实训指导书"
                    inputs:
                      prompt: "图片不少于10张"
                      docCategory: "studentHandbook"
                    status: "normal"
                    introduction: ""
                    created_at: 1742000000
                    updated_at: 1742000060
  /v1/messages:
    get:
      tags: [Dify-Compat]
      summary: 获取会话消息历史
      description: 获取某一个会话中的消息列表（Dify 兼容接口）
      parameters:
        - name: conversation_id
          in: query
          required: true
          schema:
            type: string
        - name: user
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: first_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 消息列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  limit:
                    type: integer
                  has_more:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        conversation_id:
                          type: string
                        parent_message_id:
                          type: string
                        inputs:
                          type: object
                        query:
                          type: string
                        answer:
                          type: string
                        feedback:
                          type: string
                          nullable: true
                        retriever_resources:
                          type: array
                          items: {}
                        created_at:
                          type: number
                        agent_thoughts:
                          type: array
                          items: {}
                        message_files:
                          type: array
                          items: {}
                        status:
                          type: string
                        error:
                          type: string
                          nullable: true
              example:
                limit: 20
                has_more: false
                data:
                  - id: "task_a1b2c3d4"
                    conversation_id: "test0321"
                    parent_message_id: "00000000-0000-0000-0000-000000000000"
                    inputs:
                      prompt: "图片不少于10张"
                    query: "图片不少于10张，加入3个案例"
                    answer: "文稿已生成完毕..."
                    feedback: null
                    retriever_resources: []
                    created_at: 1742000000
                    agent_thoughts: []
                    message_files: []
                    status: "normal"
                    error: null
  /v1/conversations/{id}:
    delete:
      tags: [Dify-Compat]
      summary: 删除会话
      description: 删除当前用户的某一次会话（Dify 兼容接口）
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
            example:
              user: "default"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              example:
                result: "success"
        "404":
          description: 会话不存在
  /v1/chat-messages/{task_id}/stop:
    post:
      tags: [Dify-Compat]
      summary: 停止响应
      description: 停止正在运行的流式响应（Dify 兼容接口）。task_id 是 SSE 流中的 Dify task_id。
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
          description: SSE 流中的 Dify task_id（非 message_id）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
            example:
              user: "default"
      responses:
        "200":
          description: 停止成功
          content:
            application/json:
              example:
                result: "success"
        "404":
          description: 任务不存在或已停止
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key